DjangoApi/requirements.txt
# Django 관련
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# OCR 및 이미지 처리
surya-ocr>=0.4.0
opencv-python==4.10.0.84
numpy==1.26.4
matplotlib==3.9.2
Pillow==10.4.0
PyMuPDF==1.24.10

# AI/ML
torch>=2.0.0

# 유틸리티
psutil==6.0.0
aiofiles==24.1.0
httpx==0.27.0
python-multipart==0.0.9

# 개발/테스트
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.4
